# docs/openapi/v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SimWorks API",
    "version": "0.10.2",
    "description": "REST API for SimWorks medical training platform"
  },
  "paths": {
...
    "/api/v1/simulations/": {
      "get": {
        "operationId": "api_v1_endpoints_simulations_list_simulations",
        "summary": "List user's simulations",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "default": 20,
              "description": "Max items to return",
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            },
            "required": false,
            "description": "Max items to return"
          },
          {
            "in": "query",
            "name": "cursor",
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Cursor for pagination (simulation ID)",
              "title": "Cursor"
            },
            "required": false,
            "description": "Cursor for pagination (simulation ID)"
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by status: in_progress, completed, timed_out, failed, canceled",
              "title": "Status"
            },
            "required": false,
            "description": "Filter by status: in_progress, completed, timed_out, failed, canceled"
          },
          {
            "in": "query",
            "name": "q",
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Search query",
              "title": "Q"
            },
            "required": false,
            "description": "Search query"
          },
          {
            "in": "query",
            "name": "search_messages",
            "schema": {
              "default": false,
              "description": "When true, include chat message content in search",
              "title": "Search Messages",
              "type": "boolean"
            },
            "required": false,
            "description": "When true, include chat message content in search"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_SimulationOut_"
                }
              }
            }
          }
        },
        "description": "Returns all simulations for the authenticated user, ordered by most recent first.",
        "tags": [
          "simulations"
        ],
        "security": [
          {
            "DualAuth": []
          }
        ]
      },
      "post": {
        "operationId": "api_v1_endpoints_simulations_create_simulation",
        "summary": "Create a new simulation",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulationOut"
                }
              }
            }
          }
        },
        "description": "Creates a new simulation for the authenticated user.",
        "tags": [
          "simulations"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SimulationCreate"
              }
            }
          },
          "required": true
        },
        "security": [
          {
            "DualAuth": []
          }
        ]
      }
    },
    "/api/v1/simulations/quick-create/": {
      "post": {
        "operationId": "api_v1_endpoints_simulations_quick_create_simulation",
        "summary": "Quick-create ChatLab simulation",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulationOut"
                }
              }
            }
          }
        },
        "description": "Creates a ChatLab simulation with server-generated patient identity and optional modifiers, and enqueues initial response generation.",
        "tags": [
          "simulations"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SimulationQuickCreate"
              }
            }
          },
          "required": true
        },
        "security": [
          {
            "DualAuth": []
          }
        ]
      }
    },
    "/api/v1/simulations/{simulation_id}/": {
      "get": {
        "operationId": "api_v1_endpoints_simulations_get_simulation",
        "summary": "Get simulation details",
        "parameters": [
          {
            "in": "path",
            "name": "simulation_id",
            "schema": {
              "title": "Simulation Id",
              "type": "integer"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulationOut"
                }
              }
            }
          }
        },
        "description": "Returns details for a specific simulation.",
        "tags": [
          "simulations"
        ],
        "security": [
          {
            "DualAuth": []
          }
        ]
      }
    },
    "/api/v1/simulations/{simulation_id}/end/": {
      "post": {
        "operationId": "api_v1_endpoints_simulations_end_simulation",
        "summary": "End a simulation",
        "parameters": [
          {
            "in": "path",
            "name": "simulation_id",
            "schema": {
              "title": "Simulation Id",
              "type": "integer"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulationEndResponse"
                }
              }
            }
          }
        },
        "description": "Ends an in-progress simulation and triggers feedback generation.",
        "tags": [
          "simulations"
        ],
        "security": [
          {
            "DualAuth": []
          }
        ]
      }
    },
    "/api/v1/simulations/{simulation_id}/retry-initial/": {
      "post": {
        "operationId": "api_v1_endpoints_simulations_retry_initial",
        "summary": "Retry initial simulation generation",
        "parameters": [
          {
            "in": "path",
            "name": "simulation_id",
            "schema": {
              "title": "Simulation Id",
              "type": "integer"
            },
            "required": true
          }
        ],
        "responses": {
          "202": {
            "description": "Accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulationOut"
                }
              }
            }
          }
        },
        "description": "Retries initial patient generation for a failed simulation.",
        "tags": [
          "simulations"
        ],
        "security": [
          {
            "DualAuth": []
          }
        ]
      }
    },
    "/api/v1/simulations/{simulation_id}/retry-feedback/": {
      "post": {
        "operationId": "api_v1_endpoints_simulations_retry_feedback",
        "summary": "Retry feedback generation",
        "parameters": [
          {
            "in": "path",
            "name": "simulation_id",
            "schema": {
              "title": "Simulation Id",
              "type": "integer"
            },
            "required": true
          }
        ],
        "responses": {
          "202": {
            "description": "Accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulationOut"
                }
              }
            }
          }
        },
        "description": "Retries post-simulation feedback generation.",
        "tags": [
          "simulations"
        ],
        "security": [
          {
            "DualAuth": []
          }
        ]
      }
    },
...
    "/api/v1/trainerlab/simulations/": {
      "post": {
        "operationId": "api_v1_endpoints_trainerlab_create_trainer_session",
        "summary": "Create TrainerLab simulation",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainerRunOut"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainerRunOut"
                }
              }
            }
          }
        },
        "tags": [
          "trainerlab"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrainerSessionCreateIn"
              }
            }
          },
          "required": true
        },
        "security": [
          {
            "JWTAuth": []
          }
        ]
      },
      "get": {
        "operationId": "api_v1_endpoints_trainerlab_list_trainer_sessions",
        "summary": "List TrainerLab simulations for user",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "cursor",
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            },
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_TrainerRunOut_"
                }
              }
            }
          }
        },
        "tags": [
          "trainerlab"
        ],
        "security": [
          {
            "JWTAuth": []
          }
        ]
      }
    },
    "/api/v1/trainerlab/simulations/{simulation_id}/": {
      "get": {
        "operationId": "api_v1_endpoints_trainerlab_get_trainer_session",
        "summary": "Get TrainerLab simulation",
        "parameters": [
          {
            "in": "path",
            "name": "simulation_id",
            "schema": {
              "title": "Simulation Id",
              "type": "integer"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainerRunOut"
                }
              }
            }
          }
        },
        "tags": [
          "trainerlab"
        ],
        "security": [
          {
            "JWTAuth": []
          }
        ]
      }
    },
...
    "/api/v1/trainerlab/simulations/{simulation_id}/retry-initial/": {
      "post": {
        "operationId": "api_v1_endpoints_trainerlab_retry_trainer_initial_generation",
        "summary": "Retry initial TrainerLab scenario generation",
        "parameters": [
          {
            "in": "path",
            "name": "simulation_id",
            "schema": {
              "title": "Simulation Id",
              "type": "integer"
            },
            "required": true
          }
        ],
        "responses": {
          "202": {
            "description": "Accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainerRunOut"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "trainerlab"
        ],
        "security": [
          {
            "JWTAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PaginatedResponse_SimulationOut_": {
        "properties": {
          "items": {
            "description": "List of items in this page",
            "items": {
              "$ref": "#/components/schemas/SimulationOut"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Cursor for fetching the next page, null if no more items",
            "examples": [
              "550e8400-e29b-41d4-a716-446655440000"
            ],
            "title": "Next Cursor"
          },
          "has_more": {
            "description": "Whether more items exist beyond this page",
            "title": "Has More",
            "type": "boolean"
          }
        },
        "required": [
          "items",
          "has_more"
        ],
        "title": "PaginatedResponse[SimulationOut]",
        "type": "object"
      },
      "SimulationOut": {
        "description": "Output schema for a simulation.",
        "properties": {
          "id": {
            "description": "Simulation ID",
            "title": "Id",
            "type": "integer"
          },
          "user_id": {
            "description": "Owner user ID",
            "title": "User Id",
            "type": "integer"
          },
          "start_timestamp": {
            "description": "When the simulation started",
            "format": "date-time",
            "title": "Start Timestamp",
            "type": "string"
          },
          "end_timestamp": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "When the simulation ended (null if in progress)",
            "title": "End Timestamp"
          },
          "time_limit_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Time limit in seconds (null if no limit)",
            "title": "Time Limit Seconds"
          },
          "diagnosis": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Diagnosis for this simulation",
            "title": "Diagnosis"
          },
          "chief_complaint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Chief complaint for this simulation",
            "title": "Chief Complaint"
          },
          "patient_display_name": {
            "description": "Display name for the simulated patient",
            "title": "Patient Display Name",
            "type": "string"
          },
          "patient_initials": {
            "description": "Initials for the simulated patient",
            "title": "Patient Initials",
            "type": "string"
          },
          "status": {
            "description": "Current status of the simulation",
            "enum": [
              "in_progress",
              "completed",
              "timed_out",
              "failed",
              "canceled"
            ],
            "title": "Status",
            "type": "string"
          },
          "terminal_reason_code": {
            "default": "",
            "description": "Machine-readable terminal reason code",
            "title": "Terminal Reason Code",
            "type": "string"
          },
          "terminal_reason_text": {
            "default": "",
            "description": "User-safe terminal reason text",
            "title": "Terminal Reason Text",
            "type": "string"
          },
          "terminal_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Timestamp when simulation entered terminal state",
            "title": "Terminal At"
          },
          "retryable": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "description": "Whether the failed simulation can be retried by the user",
            "title": "Retryable"
          }
        },
        "required": [
          "id",
          "user_id",
          "start_timestamp",
          "patient_display_name",
          "patient_initials",
          "status"
        ],
        "title": "SimulationOut",
        "type": "object"
      },
      "SimulationCreate": {
        "description": "Input schema for creating a simulation.",
        "properties": {
          "diagnosis": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Diagnosis for the simulation",
            "title": "Diagnosis"
          },
          "chief_complaint": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Chief complaint for the simulation",
            "title": "Chief Complaint"
          },
          "patient_full_name": {
            "description": "Full name for the simulated patient",
            "maxLength": 100,
            "minLength": 1,
            "title": "Patient Full Name",
            "type": "string"
          },
          "time_limit_seconds": {
            "anyOf": [
              {
                "maximum": 86400,
                "minimum": 60,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Time limit in seconds (null for no limit)",
            "title": "Time Limit Seconds"
          }
        },
        "required": [
          "patient_full_name"
        ],
        "title": "SimulationCreate",
        "type": "object"
      },
      "SimulationQuickCreate": {
        "description": "Input schema for one-tap ChatLab simulation creation.",
        "properties": {
          "modifiers": {
            "description": "Optional simulation modifier keys to apply.",
            "items": {
              "type": "string"
            },
            "title": "Modifiers",
            "type": "array"
          }
        },
        "title": "SimulationQuickCreate",
        "type": "object"
      },
      "SimulationEndResponse": {
        "description": "Response for ending a simulation.",
        "properties": {
          "id": {
            "description": "Simulation ID",
            "title": "Id",
            "type": "integer"
          },
          "end_timestamp": {
            "description": "When the simulation was ended",
            "format": "date-time",
            "title": "End Timestamp",
            "type": "string"
          },
          "status": {
            "const": "completed",
            "default": "completed",
            "description": "Final status",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "end_timestamp"
        ],
        "title": "SimulationEndResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "SessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sessionid"
      },
      "JWTAuth": {
        "type": "http",
        "scheme": "bearer"
      },
      "DualAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "servers": []
}